+++ src/misc/CalculatorEngineV2/Unit.js
@@ -6,9 +6,9 @@ export class Unit {
       throw new TypeError('Abstract class "Unit" cannot be instantiated directly.')
     }
 
+    // The name is used both as the binding value and the displayed text
+    // with a vue select UI object.
     if (!initObj.name) throw new Error('Please provide a unit name via initObj.name to Unit.constructor().')
     this.name = initObj.name
-    // To support vue binding with an select input, we duplictae the text as the "value"
-    this.value = this.name
   }
 }